Refactor error handling by centralizing handlers in a dedicated utility module
