+++ src/backend/app.py
@@ -1,8 +1,7 @@
 # File: app.py
 
 import os
-from flask import Flask, jsonify, g
-from werkzeug.exceptions import HTTPException
+from flask import Flask
 from flask_cors import CORS
 
 from api.home_routes import home_api_bp
@@ -10,6 +9,7 @@ from api.completions_routes import completions_api_bp
 
 from utils.logging_config import configure_logging
 from utils.env_config import list_effective_config
+from utils.error_handling import register_error_handlers
 
 
 def create_app():
@@ -20,41 +20,11 @@ def create_app():
     log_dir = os.path.join(os.path.dirname(__file__), 'logs')
     os.makedirs(log_dir, exist_ok=True)
 
-    # Distinguish HTTPException (use its code) vs generic exceptions
-    def handle_generic_exception(e):
-        app.logger.exception("Unhandled exception: %s", e)
-        cid = getattr(g, 'correlation_id', None)
-        # Allow explicit runtime misconfiguration messages to surface; mask others
-        message = 'Internal server error'
-        if isinstance(e, RuntimeError) and 'configuration incomplete' in str(e):
-            message = str(e)
-        body = {'error': message}
-        if cid:
-            body['correlation_id'] = cid
-        return jsonify(body), 500
-
-    def handle_http_exception(e: HTTPException):
-        # Provide consistent JSON error response
-        description = getattr(e, 'description', str(e))
-        cid = getattr(g, 'correlation_id', None)
-        body = {'error': description}
-        if cid:
-            body['correlation_id'] = cid
-        return jsonify(body), getattr(e, 'code', 500)
-
-    def handle_404(e: HTTPException):
-        cid = getattr(g, 'correlation_id', None)
-        body = {'error': 'Resource not found'}
-        if cid:
-            body['correlation_id'] = cid
-        return jsonify(body), 404
-
     # Configure logging (guard against duplicate handlers on reload)
     configure_logging(app)
-    # Register specific then generic
-    app.register_error_handler(404, handle_404)
-    app.register_error_handler(HTTPException, handle_http_exception)
-    app.register_error_handler(Exception, handle_generic_exception)
+
+    # Register error handlers
+    register_error_handlers(app)
 
     # Home routes
     app.register_blueprint(home_api_bp, name='home_route_direct')
